// src/app/api/month.type.ts
export enum MonthType {
    EARLY_SPRING = "early_spring",
    SPRING = "spring",
    LATE_SPRING = "late_spring",
    EARLY_SUMMER = "early_summer",
    SUMMER = "summer",
    LATE_SUMMER = "late_summer",
    EARLY_AUTUMN = "early_autumn",
    AUTUMN = "autumn",
    LATE_AUTUMN = "late_autumn",
    EARLY_WINTER = "early_winter",
    WINTER = "winter",
    LATE_WINTER = "late_winter"
}